Update views.py
In views.py, define the view named log_message (as referred to by the URL route). This view handles both HTTP GET and POST cases. In the GET case (the else: section), it just displays the form that you defined in the previous steps. In the POST case, it retrieves the data from the form into a data object (message), sets the timestamp, then saves that object at which point it's written to the database: